move feature specs into service
